test: added IT for slices (#355)
Added IT to illustrate the situation when Artipie stacks on the second request. Note, that slice should be introduced over SliceRoute and RtRulePath chain as SliceSimple with some response. If SliceSimple is replaced with a proper separate Slice implementation (as AuthTokenSlice.java in conde for example) everything works properly.

Issue: #356